### 2025-05-12 #8189 @editorialbot
[PRE REVIEW]: Scikit-Longitudinal: A Machine Learning Library for Longitudinal Classification in Python
**Submitting author:** @simonprovost (<a href="http://orcid.org/0000-0001-8402-5464">Simon Provost</a>)
**Repository:** https://github.com/simonprovost/scikit-longitudinal
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.0.7
