### 2025-05-12 #8189 @simonprovost
# Questions for prospective reviewers

- `The word count for paper.md is 1295`. ––  We had an internal discussion with my professor (Prof. Alex Freitas) about how we needed the two hundred extra characters for terminology and credits.  Can we go above the 1,000 limit if it is reasonable, or is it strictly 1k and needs to be updated?  If so, what are your suggestions?

 - The paper's author list includes `Prof. Alex Freitas`.  Prof. Alex Freitas did not directly contribute to the code aspect directly but rather indirectly. Its design, architecture, and the majority of the primitives are the result of his main supervision with previous and current researchers (including myself). This will be a paper for my Ph.D. and thus cannot be done without my supervisor. Is this going to be problematic?

I would like to thank you very much my prospective reviewers for your valuable time with all of that!  I cannot wait to hear from any of you!

Cheers,
