+++ experiments/experiment_launchers/Lexicographical Trees paper/deep_forest_HPO.py
@@ -1,0 +1,372 @@
+import argparse
+import os
+from typing import List
+
+import numpy as np
+import pandas as pd
+from imblearn.metrics import geometric_mean_score
+from imblearn.over_sampling import SMOTE
+from imblearn.pipeline import Pipeline
+from imblearn.under_sampling import RandomUnderSampler
+from scikit_longitudinal.data_preparation import LongitudinalDataset
+from sklearn.base import BaseEstimator, ClassifierMixin
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.metrics import make_scorer
+from sklearn.model_selection import StratifiedKFold, GridSearchCV
+
+from experiments.experiment_engine import ExperimentEngine
+from gama.utilities.longitudinal_config_space import CascadeForestClassifier
+
+
+class CustomDeepForestClassifier(BaseEstimator, ClassifierMixin):
+    def __init__(self, rf_params=None, random_state=None, max_layers=5):
+        """
+        Custom Deep Forest Classifier with two Random Forests and two Completely Random Forests.
+        Allows hyperparameter tuning for the Random Forests.
+
+        Parameters:
+        - rf_params: dict, optional
+            Parameters for the two main Random Forests.
+        - random_state: int, optional
+            Seed for reproducibility.
+        - max_layers: int, optional
+            Maximum layers in the CascadeForestClassifier.
+        """
+        self.rf_params = rf_params if rf_params else {}
+        self.random_state = random_state
+        self.max_layers = max_layers
+        self._cascade = None
+
+    def _initialize_estimators(self):
+        # Two Random Forests with customizable hyperparameters
+        rf1 = RandomForestClassifier(**self.rf_params, random_state=self.random_state)
+        rf2 = RandomForestClassifier(**self.rf_params, random_state=self.random_state)
+        # Two Completely Random Forests (max_features set to 1)
+        crf1 = RandomForestClassifier(max_features=1, random_state=self.random_state)
+        crf2 = RandomForestClassifier(max_features=1, random_state=self.random_state)
+        return [rf1, rf2, crf1, crf2]
+
+    def fit(self, X, y):
+        """
+        Fit the custom deep forest model on training data.
+
+        Parameters:
+        - X: array-like of shape (n_samples, n_features)
+            Training input samples.
+        - y: array-like of shape (n_samples,)
+            Target values (class labels).
+        """
+        self._cascade = CascadeForestClassifier(random_state=self.random_state, max_layers=self.max_layers)
+        self._cascade.set_estimator(self._initialize_estimators(), n_splits=2)
+        self._cascade.fit(X, y)
+        self.classes_ = np.unique(y)
+        return self
+
+    def predict(self, X):
+        """
+        Predict class labels for samples in X.
+
+        Parameters:
+        - X: array-like of shape (n_samples, n_features)
+            Input samples.
+
+        Returns:
+        - y_pred: array of shape (n_samples,)
+            Predicted class labels.
+        """
+        return self._cascade.predict(X)
+
+    def predict_proba(self, X):
+        """
+        Predict class probabilities for samples in X.
+
+        Parameters:
+        - X: array-like of shape (n_samples, n_features)
+            Input samples.
+
+        Returns:
+        - y_proba: array of shape (n_samples, n_classes)
+            Predicted class probabilities.
+        """
+        return self._cascade.predict_proba(X)
+
+
+def _load_dataset(
+        dataset_file_path: str,
+        target_column: str,
+        random_state: int,
+        remove_target_waves: bool = True,
+        setup_features_group="elsa"
+) -> (pd.DataFrame, pd.Series, List[List[int]], List[int], List[str]):
+    """
+    Loads the dataset from the specified file path, processes it, and prepares it for use in the experiment.
+
+    Args:
[…]
