+++ mkdocs.yml
@@ -159,5 +159,6 @@ nav:
       - "Longitudinal Pipeline": API/pipeline.md
   - "Examples": examples/index.md
   - "Contributing": contribution.md
+  - "Experiments": experiments.md
   - "Release History": https://github.com/simonprovost/scikit-longitudinal/releases
   - "FAQ": faq.md

[…]
