### 2025-06-25 #8481 @editorialbot
[REVIEW]: Scikit-Longitudinal: A Machine Learning Library for Longitudinal Classification in Python
**Submitting author:** @simonprovost (<a href="http://orcid.org/0000-0001-8402-5464">Simon Provost</a>)
**Repository:** https://github.com/simonprovost/scikit-longitudinal
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.0.9
