### 2025-06-26 d78c84c
refactor(docs): improve import & docstrings' examples
