### 2025-06-26 08733c4
refactor(docs): improve docs wordings
