+++ docs/getting-started.md
@@ -3,34 +3,64 @@ hide:
   - navigation
 ---
 
-## 💡 Getting Started with `Scikit-longitudinal`
+# 💡 `Scikit-longitudinal`, in a nutshell!
+# 💡 `Scikit-longitudinal`, in a nutshell!
 
-Longitudinal datasets contain information about the same cohort of individuals (instances) over time, with the same set
-of features (variables) repeatedly measured across different time points (also called `waves`)[^1][^2][^3].
+Biomedical research often uses longitudinal data with repeated measurements of variables across time (e.g. cholesterol
+measured across time), which is challenging for standard machine learning algorithms due to intrinsic temporal
+dependencies. 
 
-`Scikit-longitudinal` (Sklong) is a machine learning library designed to analyse longitudinal data, also known as _Panel
-data_ in certain fields. Today, Sklong focuses on Longitudinal Machine Learning _Classification_ tasks. It offers tools
-and models for `processing`, `analysing`, and `classifying` longitudinal tabular data, with a user-friendly 
-interface that integrates with the `Scikit-learn` ecosystem.
+`Scikit-longitudinal` (abbreviated `Sklong`, pronounced /ˌɛs keɪ ˈlɒŋ/ or "Ess-kay-long" and /ˌsaɪ kɪt
+ˌlɒndʒɪˈtjuːdɪnəl/ or "Sky-kit lon-ji-TOO-din-ul") is a machine learning library helping out in anlysing longitudinal 
+data, also known as _panel data_ in some fields. `Sklong` specialises in longitudinal machine learning _classification_ tasks,
+offering user-friendly tools for `processing`, `analyzing`, and `classifying` longitudinal tabular data, seamlessly
+integrating with the `Scikit-learn` ecosystem.
 
-Let's first jump into the installation, and then we will explore the first steps with code and some common questions.
-We then suggest to explore the API reference and the examples to get a better understanding of the library.
+Note that while Longitudinal datasets have a temporal component, other types of datasets, such as time series,
+also have a temporal component but are not considered longitudinal datasets. Time series data typically involves
+a single variable measured at regular intervals over time, while longitudinal datasets involve multiple variables
+measured across the same cohort of individuals at different time points. More is discussed in the [FAQ](https://scikit-longitudinal.readthedocs.io/latest//faq/).
+However, I would like to highlight that time points are therefore considered as `waves` in `Sklong` [^1][^2][^3].
+
+To start your Longitudinal Machine Learning journey with `Sklong`, you first will have to install the library.
 
 ---
 
 ## 🛠️ Installation
 
 !!! warning "Operating System Support"
-    `Scikit-longitudinal` is currently supported on `OSX` (MacOS) and `Linux`. `Windows` users should use `notebooks` or `Docker` with a `Linux` (
-    Ubuntu) distribution due to limitations with a dependency library. For more details, see
-    the [Scikit-Lexicographical-Trees GitHub repository](https://github.com/simonprovost/scikit-lexicographical-trees). Feel
-    free to contribute a Windows-based wheel to unlock this potential!
+    `Scikit-longitudinal` is currently supported on `OSX` (MacOS) and `Linux`. 
+    `Windows` users should use `notebooks` or `Docker` with a `Linux` (
+    Ubuntu) distribution due to limitations with a dependency library. For more details, open an issue, I would be 
+    happy to discuss this out further.
 
-### Installation Methods
+!!! important "Python Version Compatibility"
+    `Scikit-longitudinal` is currently compatible with Python versions `3.9` and `3.10`.
+    Ensure you have one of these versions installed before proceeding with the installation.
+
[…]
