### 2025-06-26 55427b1
refactor(docs): add new-style home
