### 2025-06-27 5ac0a77
refactor(docs): emphasise version compatibility
