### 2025-06-27 10e8265
release-pypi: 0.0.7->0.0.8, update pypi readme [cd build]
