### 2025-06-27 #8481 @espottesmith
aFrom [the JOSS documentation](https://joss.readthedocs.io/en/latest/paper.html): "The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper."

Your paper is about ~1300 words long. I would not consider that "significantly longer". However, I agree with @blengerich that certain sections of the paper can be tightened. It's unclear to me that this paper really needs to be more than 1000 words long.
