### 2025-06-27 54f5e5a
refactor(paper): tighten paper given word limit @ JOSS
