+++ paper/paper.md
@@ -32,7 +32,7 @@ working with longitudinal datasets.
 # Summary
 
 `Scikit-Longitudinal` is an open-source Python library that enhances machine learning for longitudinal data
-classification and integrates flawlessly with the `Scikit-Learn` environment [@pedregosa2011scikit].
+classification and integrates with the `Scikit-Learn` environment [@pedregosa2011scikit].
 
 Longitudinal data, which consists of repeated measurements of variables across time points (referred to as
 `waves` [@ribeiro2019mini]), is extensively utilised in fields such as medicine, social sciences, and economics. Unlike
@@ -53,7 +53,8 @@ processing [@kelloway2012longitudinal].
 - **Preprocessing**: Longitudinal-data-aware feature selection primitives, such
   as
   `CFS-Per-Group` [@pomsuwan2017feature],
-  utilising the temporal information in the data to proceed with feature selection techniques (see feature selection review in [@theng2024feature]).
+  utilising the temporal information in the data to proceed with feature selection techniques (see feature selection
+  review in [@theng2024feature]).
 
 - **Estimators**: Longitudinal-data-aware classifiers [@kotsiantis2007supervised] [@ribeiro2019mini],
   such as
@@ -72,8 +73,8 @@ longitudinal data analysis.
 
 Finally, `Scikit-Longitudinal` can also be abbreviated `Sklong`, and is available
 on [PyPI](https://pypi.org/project/Scikit-longitudinal/). Feel free to explore the
-[official documentation](https://scikit-longitudinal.readthedocs.io/latest/getting-started/) for various
-installation methods, including locally, via `Colab`, `Marimo`, and others.
+[official documentation](https://scikit-longitudinal.readthedocs.io/latest/) for various
+installation methods, including locally, via `Jupyter`, `Colab`, and others.
 
 # Longitudinal Classification
 
@@ -98,77 +99,51 @@ different features directly. Note that, in the type of longitudinal classificati
 the features are longitudinal, but the class variable is not; i.e.,
 the goal is to predict the class label of an instance at a single time point (usually the most recent wave).
 
-## Approaches to Longitudinal Classification
-
-There are two broad approaches for coping with longitudinal data [@ribeiro2019mini]:
-
-1. **Data Transformation**  
-   This approach involves preprocessing methods that convert longitudinal data into a standard, "flattened"
-   non-longitudinal format. This transformation enables the use of any standard, non-longitudinal classification
-   algorithm on the data. However, this approach may result in the loss of relevant information regarding how a
-   feature's values change over time.
-
-2. **Algorithm Adaptation**  
-   This approach entails modifying classification algorithms to directly handle temporal dependencies inherent in
-   longitudinal datasets. This preserves the temporal dynamics of the data but may require more complex modelling
-   techniques.
+There are two broad approaches for coping with longitudinal data [@ribeiro2019mini]: (1) **Data Transformation**:
+this approach involves preprocessing methods that convert longitudinal data into a standard, "flattened"
+non-longitudinal format. This transformation enables the use of any standard, non-longitudinal classification
+algorithm on the data. However, this approach may result in the loss of relevant information regarding how a
+feature's values change over time. (2) **Algorithm Adaptation**: this approach entails modifying classification
+algorithms to directly handle temporal dependencies inherent in
+longitudinal datasets. This preserves the temporal dynamics of the data but may require more complex tooling.
 
 # Statement of Need
 
-The temporal nature of longitudinal data poses significant challenges for conventional machine learning techniques,
-which typically assume independence among observations. Standard libraries such as `Scikit-learn` do not inherently
-support longitudinal data, resulting in inefficient or inaccurate analyses. Consequently, there is a pressing need for a
[…]
