### 2025-06-27 #8481 @simonprovost
Thanks @espottesmith and @blengerich for pushing this through! Really appreciated. I've tightened the paper and remove the story beyond `Sklong`. Can be seen via this commit: https://github.com/simonprovost/scikit-longitudinal/commit/54f5e5aa507034a6c3ef4631e97722f855dab172

Are you giving your green flag go @espottesmith or may you have other suggestions?

Cheers,
