### 2025-06-27 simonprovost/scikit-longitudinal#59 @blengerich
[JOSS] Installation
I am trying to install on Mac. I have `Python 3.9.18` running, and created a fresh venv. Running `pip install scikit-longitudinal`, I get an error trying to install `deep-forest`.

> pip install scikit-longitudinal
> Collecting scikit-longitudinal
>   Downloading scikit_longitudinal-0.0.8-py3-none-any.whl.metadata (6.4 kB)
> Collecting scipy>=1.5.0 (from scikit-longitudinal)
>   Using cached scipy-1.13.1-cp39-cp39-macosx_12_0_arm64.whl.metadata (60 kB)
> Collecting pandas<2.0.0,>=1.5.3 (from scikit-longitudinal)
>   Downloading pandas-1.5.3-cp39-cp39-macosx_11_0_arm64.whl.metadata (11 kB)
> Collecting matplotlib<4.0.0,>=3.7.0 (from scikit-longitudinal)
>   Downloading matplotlib-3.9.4-cp39-cp39-macosx_11_0_arm64.whl.metadata (11 kB)
> Collecting jupyter<2.0.0,>=1.0.0 (from scikit-longitudinal)
>   Downloading jupyter-1.1.1-py2.py3-none-any.whl.metadata (2.0 kB)
> Collecting overrides<8.0.0,>=7.3.1 (from scikit-longitudinal)
>   Using cached overrides-7.7.0-py3-none-any.whl.metadata (5.8 kB)
> Collecting statsmodels<1.0.0,>=0.14.0 (from scikit-longitudinal)
>   Downloading statsmodels-0.14.4-cp39-cp39-macosx_11_0_arm64.whl.metadata (9.2 kB)
> Collecting numpy==1.23.3 (from scikit-longitudinal)
>   Downloading numpy-1.23.3-cp39-cp39-macosx_11_0_arm64.whl.metadata (2.3 kB)
> Collecting graphviz<1.0.0,>=0.20.1 (from scikit-longitudinal)
>   Downloading graphviz-0.21-py3-none-any.whl.metadata (12 kB)
> Collecting liac-arff>=2.2.2 (from scikit-longitudinal)
>   Downloading liac-arff-2.5.0.tar.gz (13 kB)
>   Installing build dependencies ... done
>   Getting requirements to build wheel ... done
>   Preparing metadata (pyproject.toml) ... done
> Collecting threadpoolctl<4.0.0,>=3.1.0 (from scikit-longitudinal)
>   Downloading threadpoolctl-3.6.0-py3-none-any.whl.metadata (13 kB)
> Collecting stopit>=1.1.2 (from scikit-longitudinal)
>   Using cached stopit-1.1.2-py3-none-any.whl
> Collecting rich>=13.6.0 (from scikit-longitudinal)
>   Downloading rich-14.0.0-py3-none-any.whl.metadata (18 kB)
> Collecting joblib>=0.11 (from scikit-longitudinal)
>   Downloading joblib-1.5.1-py3-none-any.whl.metadata (5.6 kB)
> INFO: pip is looking at multiple versions of scikit-longitudinal to determine which version is compatible with other requirements. This could take a while.
> Collecting scikit-longitudinal
>   Downloading Scikit_longitudinal-0.0.7-py3-none-any.whl.metadata (5.8 kB)
>   Downloading scikit_longitudinal-0.0.6-py3-none-any.whl.metadata (11 kB)
> Collecting ray<3.0.0,>=2.3.1 (from scikit-longitudinal)
>   Downloading ray-2.47.1-cp39-cp39-macosx_11_0_arm64.whl.metadata (20 kB)
> Collecting scikit-longitudinal
>   Downloading scikit_longitudinal-0.0.5-py3-none-any.whl.metadata (11 kB)
>   Downloading scikit_longitudinal-0.0.4-py3-none-any.whl.metadata (11 kB)
> Collecting arff<1.0,>=0.9 (from scikit-longitudinal)
>   Downloading arff-0.9.tar.gz (4.7 kB)
>   Installing build dependencies ... done
>   Getting requirements to build wheel ... done
>   Preparing metadata (pyproject.toml) ... done
> Collecting configspace<1.0.0,>=0.7.1 (from scikit-longitudinal)
>   Downloading ConfigSpace-0.7.1-cp39-cp39-macosx_10_9_universal2.whl.metadata (7.2 kB)
> Collecting scikit-longitudinal
>   Downloading scikit_longitudinal-0.0.2-py3-none-any.whl.metadata (8.5 kB)
> ERROR: Cannot install scikit-longitudinal==0.0.2, scikit-longitudinal==0.0.4, scikit-longitudinal==0.0.5, scikit-longitudinal==0.0.6, scikit-longitudinal==0.0.7 and scikit-longitudinal==0.0.8 because these package versions have conflicting dependencies.
>
> The conflict is caused by:
>     scikit-longitudinal 0.0.8 depends on deep-forest>=0.1.7
>     scikit-longitudinal 0.0.7 depends on deep-forest>=0.1.7
>     scikit-longitudinal 0.0.6 depends on deep-forest>=0.1.7
>     scikit-longitudinal 0.0.5 depends on deep-forest>=0.1.7
>     scikit-longitudinal 0.0.4 depends on deep-forest>=0.1.7
>     scikit-longitudinal 0.0.2 depends on deep-forest>=0.1.7
>
> To fix this you could try to:
> 1. loosen the range of package versions you've specified
> 2. remove package versions to allow pip to attempt to solve the dependency conflict
>
> ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
>
