### 2025-06-27 e142440
refactor(docs): add troubleshooting install section
