### 2025-06-27 simonprovost/scikit-longitudinal#59 @simonprovost
@blengerich, these days I use `UV`, but for the sake of this current issue, here is a reproduction in video with my machine following the troubleshooting previously shared:

https://github.com/user-attachments/assets/0fb1f533-72ee-4a95-a5fb-dba9172ab751
