### 2025-07-03 #8481 @blengerich
Hi @simonprovost , I have updated review with installation, etc. It is unfortunate that the dependency limits the installation so drastically; any way you can improve that will likely improve the usability of your package. That being said, you've properly documented installation requirements.

The final item on my checklist is the writing quality. Given that your draft is now 1101 words, I think it should be possible to have a version <1000 words and not need to loosen any editorial guidelines.
