### 2025-07-03 simonprovost/scikit-longitudinal#59 @simonprovost
I have two options that I wish to explore in September as this summer I'll focus my thesis:

1. Fork and try to upgrade `Deep Forest` to 3.10 myself?
2. Make optional the `Lexicographical Deep Forest` primitive via `sckit-longitudinal[lexicographical_deep_forest]` which would therefore require downgrading versions by the user but quite complex given the potential end users.

Optimistically speaking, option is the best, for everyone :) But I have received no answer on the github issue I pinned earlier with their authors about what are the risks to increase to py3.10 :(

I 100% agree that this is so very much unfortunate, sadly it is a nice algorithm!

Cheers,
