### 2025-07-03 #8481 @simonprovost
Hi @blengerich,

> I have updated review with installation, etc. It is unfortunate that the dependency limits the installation so drastically; any way you can improve that will likely improve the usability of your package. That being said, you've properly documented installation requirements.

100% agree; it's in my head to explore after the summer if things go well on writing the thesis. It definitely is an important point because in the long term it could even be more dramatic.

TLDR for editor and all; It all works but could be quite tricky for newcomers. Therefore I proposed two exploration after the summer, available at https://github.com/simonprovost/scikit-longitudinal/issues/59#issuecomment-3033536369

> The final item on my checklist is the writing quality. Given that your draft is now 1101 words, I think it should be possible to have a version <1000 words and not need to loosen any editorial guidelines.

I'll try to explore that tomorrow; may I see what you'd be choosing as to perhaps "too long" to refactor? If you do not answer prior I start, that's fine ^^

Thanks for your time!

Cheers
