### 2025-07-04 c137dda
refactor(paper): tighten paper to <1000 words @ JOSS review
