+++ paper/paper.md
@@ -25,8 +25,8 @@ bibliography: paper.bib
 Longitudinal data, characterised by repeated measurements of variables over time, presents unique challenges and
 opportunities in
 machine learning. This paper introduces `Scikit-Longitudinal`, a Python library designed to address these challenges by
-providing a comprehensive set of tools for longitudinal data classification. Built to integrate flawlessly with the
-widely-used `Scikit-Learn` library, `Scikit-Longitudinal` offers a robust solution for researchers and practitioners
+providing a comprehensive set of tools for longitudinal data classification. Built to integrate with the
+popular `Scikit-Learn` library, `Scikit-Longitudinal` offers a robust solution for researchers and practitioners
 working with longitudinal datasets.
 
 # Summary
@@ -35,7 +35,7 @@ working with longitudinal datasets.
 classification and integrates with the `Scikit-Learn` environment [@pedregosa2011scikit].
 
 Longitudinal data, which consists of repeated measurements of variables across time points (referred to as
-`waves` [@ribeiro2019mini]), is extensively utilised in fields such as medicine, social sciences, and economics. Unlike
+`waves` [@ribeiro2019mini]), is extensively utilised in fields such as medicine, and social sciences. Unlike
 standard tabular datasets, it contains temporal relationships that necessitate specialised
 processing [@kelloway2012longitudinal].
 
@@ -68,13 +68,12 @@ In total, the library implements 1 data preparation method, 4 data transformatio
 `LexicoRandomForestClassifier` and `NestedTreesClassifier` methods).
 
 `Scikit-Longitudinal` emphasises highly-typed, Pythonic code, with substantial test coverage (over 88%) and
-comprehensive documentation (over 72%), ensuring reliability and extensibility for researchers and practitioners in
-longitudinal data analysis.
+comprehensive documentation (over 72%).
 
 Finally, `Scikit-Longitudinal` can also be abbreviated `Sklong`, and is available
 on [PyPI](https://pypi.org/project/Scikit-longitudinal/). Feel free to explore the
 [official documentation](https://scikit-longitudinal.readthedocs.io/latest/) for various
-installation methods, including locally, via `Jupyter`, `Colab`, and others.
+installation methods.
 
 # Longitudinal Classification
 
@@ -97,7 +96,7 @@ standard machine learning algorithms, or handle the temporal dependencies betwee
 between
 different features directly. Note that, in the type of longitudinal classification task for which `Sklong` was designed,
 the features are longitudinal, but the class variable is not; i.e.,
-the goal is to predict the class label of an instance at a single time point (usually the most recent wave).
+the goal is to predict the class label of an instance at a single time point (usually the last wave).
 
 There are two broad approaches for coping with longitudinal data [@ribeiro2019mini]: (1) **Data Transformation**:
 this approach involves preprocessing methods that convert longitudinal data into a standard, "flattened"
@@ -109,41 +108,33 @@ longitudinal datasets. This preserves the temporal dynamics of the data but may 
 
 # Statement of Need
 
-The temporal aspect of longitudinal data complicates conventional machine learning techniques, which presume observation
-independence. Standard libraries, such as `Scikit-learn`, lack support for longitudinal data, resulting in inefficient
-or inaccurate analysis.
-To the best of our knowledge, there is no package in the `Scikit-learn` ecosystem that offers a straightforward solution
-for longitudinal classification.
+To the best of our knowledge, no package in the `Scikit-learn` ecosystem provides an easy solution for longitudinal
+classification.
+Standard Python libraries, such as `Scikit-learn` itself, lack support for longitudinal data, leading to inefficient and
+inaccurate analysis. `R` includes
[…]
