### 2025-07-04 #8481 @simonprovost
> My review is now completed [@espottesmith](https://github.com/espottesmith) [@simonprovost](https://github.com/simonprovost) .

Thanks so much @blengerich ! It was very informative :)

@espottesmith Paper is now 1000 words sharp ! What's next please?
