### 2025-07-08 simonprovost/scikit-longitudinal#60 @simonprovost
Hi @TahiriNadia,

Thank you so much for taking the time to read your peer review section of the JOSS paper :) !

Your issue is due to the fact that you are using Python `3.12`; However, it is mentioned as not possible for the time being, see the screenshot below.

In that instance, having different formal Python versions on the same system can be a nightmare at times —At least, from experience.  That is why we advocate utilising `uv` from Astral, which can download different Python versions in a more compartmentalised manner for you. See here https://scikit-longitudinal.readthedocs.io/latest/getting-started/#__tabbed_1_3 yet if Jupyter is not of interest simply follow the `uv python install 3.9` then `uv python pin 3.9` followed by `uv add scikit-longitudinal` and you should be good.

Please, If you know how to do it without `uv` (Pyenv, Conda, etc.), that is entirely fine, `uv` is simply a recommendation from our side of things given it is a SOTA package manager nowadays. The only thing you need is: to use a Python version `>3.9+ <3.10`.

Following that, if you are on MacOS, please be aware that there are known tricks with Apple Silicon; while this is noted in the documentation, I will link to the Github issue that demonstrates it:  https://github.com/simonprovost/scikit-longitudinal/issues/59

Hoping it all goes through, let me know I am always fast answering — At least, as fast as possible :) !

Cheers,
