### 2025-07-10 simonprovost/scikit-longitudinal#60 @simonprovost
Hi @TahiriNadia,

Have you had the chance to look into this :) ?

Cheers,
