### 2025-07-10 simonprovost/scikit-longitudinal#60 @TahiriNadia
Hi @TahiriNadia,

For sure, I will work on it in the coming days.

Cheers,
