### 2025-07-10 simonprovost/scikit-longitudinal#60 @simonprovost
Thanks so muchhh!
