### 2025-07-14 simonprovost/scikit-longitudinal#61 @TahiriNadia
[JOSS] 💡 Feature Request - tutorial
### 🤔 What's the Problem?

Currently, there is no complete tutorial or clearly accessible example in the repository that demonstrates how to use `scikit-longitudinal` in a typical analysis workflow. If an example exists, it is difficult to find or too generic—it requires copying and pasting code and finding an appropriate dataset separately. For new users, it's not immediately clear how to go from a raw longitudinal dataset to a fitted model using the API.
