### 2025-07-14 simonprovost/scikit-longitudinal#62 @TahiriNadia
[JOSS] 💡 Feature Request - Workflow picture
# 🤔 What's the Problem?

To support a clearer understanding of the proposed feature, a workflow illustration or structured outline may be included. Such a representation can assist in evaluating how the feature integrates with the existing `Scikit-longitudinal` architecture.
