### 2025-07-14 simonprovost/scikit-longitudinal#62 @simonprovost
Hi @TahiriNadia,

Thanks much for this extra FR! :)

I would like to kindly vote against this because this is not _one and only one_ algorithm that is proposed — as in, an algorithm 'proposed as part of a paper — throughout Sklong.

There is not only one single new algorithm for instance that we would need to visually represent.  But all of them. I believe their respective papers should do so (`Publication` tab shows the primitives' various papers).

As a ressult, Sklong is a hub for several various primitives, each with a unique design but comparable goal (Longitudinal classification).  To add to that, Sklearn does not provide such a representation, does it?

Now, I am delighted to be proven wrong, but may I please ask what you had in mind with this illustration?  If it was a figure comparable to how we usually visually express an algorithm/approach/methodology in a paper, I feel it is currently impracticable plus the various papers re implemented have their own visual/algorithmic details there.  If you mean something altogether different, I am all ears :)

 Cheers,
