### 2025-07-14 #8481 @TahiriNadia
`Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?`
Avoid `.s` lines 58, 66 and 90.
