### 2025-07-14 #8481 @simonprovost
> `Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?` Avoid `.s` lines 58, 66 and 90.

Hi @TahiriNadia,

Very appreciated for your feedback on the paper’s writing quality.

However, the use of **'s** in these lines (`feature's values` in lines `58` and `66`, and `Scikit-learn's Cython trees` in line `⁣`)—AFAIR—sounds grammatically correct, as it forms _possessives for singular nouns according to standard rules_ (reference: https://owl.purdue.edu/owl/general_writing/punctuation/apostrophe_introduction.html).

 For example, at line 58, the phrase `temporal dependencies between a feature's values across time` refers to the values of a single feature tracked over time, proving that `feature` is singular and requires **'s** for possession.  Similarly, in line 66, `feature's values` refers to one feature's temporal data, and in line 90, `Scikit-learn` is a solitary proper noun that owns the Cython trees, both of which correctly use **'s**, IMHO.

 I am pleased to be mistaken, but may you please let me know how?

 Cheers
