### 2025-07-14 #8481 @simonprovost
> Issues list:
>  [...]

Thanks so so much for your time on these @TahiriNadia, I've finished my round and waiting for answers whenever you've got time, obviously. 🙏

| Issue | Status | Summary |
|-------|--------|---------|
| [[JOSS] 🚨 BUG - codespace](https://github.com/simonprovost/scikit-longitudinal/issues/60) | ✅ | Installation failed due to wrong Python version, solved by guiding to right part of the doc. |
| [[JOSS] 💡 Feature Request - Workflow picture](https://github.com/simonprovost/scikit-longitudinal/issues/62) | ⏳|  Request for a workflow illustration or structured outline to clarify integration with existing scikit-longitudinal architecture. |
| [[JOSS] 💡 Feature Request - tutorial](https://github.com/simonprovost/scikit-longitudinal/issues/61) | 💡 | Request for a complete tutorial or easily accessible example showing the full analysis workflow from raw longitudinal dataset to fitted model.

Cheers
