### 2025-07-15 simonprovost/scikit-longitudinal#61 @simonprovost
I think it was somehow shown in the `Temporal Dependency` tab but maybe not enough, let me work on that tomorrow or next days, I'll tag you back here for your approval, thanks so so much @TahiriNadia !
