### 2025-07-15 simonprovost/scikit-longitudinal#61 @TahiriNadia
Yes, I did notice the demonstration in the Temporal Dependency tab, and it’s a good starting point — but in my opinion, it’s not quite sufficient yet to fully showcase how sklong handles real longitudinal structures. I think expanding it a bit further, perhaps with a more guided example, could help clarify things. Just a suggestion, of course. Thank you, @simonprovost!
