### 2025-07-15 simonprovost/scikit-longitudinal#61 @simonprovost
Good suggestions, there are never bad suggestions from an early age of a software anyway :)

Will work on that and will tag you back here then,

Cheerrs
