### 2025-07-21 #8481 @simonprovost
@espottesmith Sorry not to keep you updated! Prof. @TahiriNadia requested a couple of modifications we agreed upon, I am working on it today (doc-wise mainly)

Hopefully all is good afterward
