### 2025-07-21 61caecb
core: add glightbox dev dep for mkdocs
