+++ pyproject.toml
@@ -57,6 +57,7 @@ dev = [
     "pymdown-extensions>=10.14.3",
     "mkdocs-material[emoji,imaging]>=9.6.11",
     "mkdocs-document-dates>=3.1.5",
+    "mkdocs-glightbox>=0.4.0",
 ]
 
 [urls]
