### 2025-07-21 582725a
refactor(docs): adapt temporal dep link
