+++ scikit_longitudinal/estimators/trees/lexicographical/lexico_decision_tree_regressor.py
@@ -42,7 +42,7 @@ class LexicoDecisionTreeRegressor(DecisionTreeRegressor):
         use the primitives that follow.
 
         To see more, we highly recommend visiting the `Temporal Dependency` page in the documentation.
-        [Temporal Dependency Guide :fontawesome-solid-timeline:](https://scikit-longitudinal.readthedocs.io/latest//temporal_dependency/){ .md-button }
+        [Temporal Dependency Guide :fontawesome-solid-timeline:](https://scikit-longitudinal.readthedocs.io/latest/tutorials/temporal_dependency/){ .md-button }
 
     Args:
         threshold_gain (float, default=0.0015):
