### 2025-07-21 simonprovost/scikit-longitudinal#61 @simonprovost
Hi, @TahiriNadia!

I hope everything is well with you, and thank you for your excellent suggestions; I greatly liked them!

Even was taken into account ——as agreed anyway——, changes can be seen in PR #63.  New documentation can be seen in action at https://scikit-longitudinal.readthedocs.io/latest/.

Should we consider this thread completed?  Shall we proceed with the JOSS review:)

 Cheers,
