### 2025-07-22 simonprovost/scikit-longitudinal#61 @TahiriNadia
Hi @simonprovost! 😊

Thanks a lot for the quick updates 🔥  — everything looks great on my end, and the changes are well done 💃. Let us consider this thread completed and move forward with the JOSS review!

Cheers,
