### 2025-07-22 #8481 @TahiriNadia
> All right, new documentation's out following [@TahiriNadia](https://github.com/TahiriNadia)'s kind peer review. See @ https://scikit-longitudinal.readthedocs.io/latest/
>
> Anything else [@TahiriNadia](https://github.com/TahiriNadia) you'd like to discuss about?
>
> Thank you so so much both [@TahiriNadia](https://github.com/TahiriNadia) and [@blengerich](https://github.com/blengerich) for your kind time!
>
> Cheers

The latest modifications from @SimonProvost look spot on — thanks for the careful work! 🔧✅
