### 2025-07-22 #8481 @simonprovost
All right, on it now will be done in 5" — Will make a full new release note after the JOSS paper published, just FYI (will cc-credit both of my reviewer and you as my editor as well)
