### 2025-07-22 #8481 @simonprovost
@espottesmith

- Commit for new release (`0.0.9`): https://github.com/simonprovost/scikit-longitudinal/commit/deb3ef5ae3c750fc461df69b26c1d3ac5d20f78a

- Avail on PyPi as `0.0.9`: https://pypi.org/project/Scikit-longitudinal/0.0.9/#history

Once more, a brand new release will be available after the paper published with tag 0.0.9
