### 2025-07-25 #8481 @espottesmith
Quick question @simonprovost: why do you have two licenses listed on Zenodo (MIT and a CC license), but only one on GitHub (MIT)? What part of the code is CC-BY, and what part is covered by MIT?
