### 2025-07-28 #8481 @simonprovost
> Quick question [@simonprovost](https://github.com/simonprovost): why do you have two licenses listed on Zenodo (MIT and a CC license), but only one on GitHub (MIT)? What part of the code is CC-BY, and what part is covered by MIT?

Hi @espottesmith  (answering more formally)

Sorry about the two licenses confusion. I will chalk this up to the fact that it is my first time with Zenodo ^^ However, the issue was primarily caused by Zenodo's default license (`CC`).   I just added `MIT` as it is Sklong's main one.

Anyway, thanks much for pointing that out! Just removed the `CC` on Zenodo because it is anyway not present in the SKlong repository (only `MIT` is).

May you kindly please let me know if you need anything else from me? I am a little bit hoping we may have the chance to advance things further as much as possible before meeting with my Ph.D. panel interviewer on Wednesday; I am not expecting anything, of course, and will wait the time it needs to be waited. no worries about that; just fingers crossed. :)

Anyway, note that I am available at any single moment of the day for the next few days; so feel free to request changes on the fly if necessary.

Wishing you a very nice start to the week!

 Cheers,
