### 2025-07-28 #8481 @espottesmith
Please address the reference issues listed above.

### Paper Comments:
1. "is extensively utilised in fields such as medicine, and social science" <-- there should be no comma here, since it's a list of only two entries.
2. "Unlike standard tabular datasets, it contains temporal relationships that necessitate specialised processing" <-- "it" is somewhat ambiguous here
3. "(i.e. ignoring time indices)" <-- there should be a comma after "i.e."
4. "stand alone" should be "stand-alone"
5. "Scikit-Longitudinal can also be abbreviated Sklong, and is available on PyPI." <-- no comma is needed here. Also, the abbreviation should probably be introduced at the beginning of your manuscript and used throughout.
6. There's quite a bit of repetition w/ respect to the two approaches for dealing with longitudinal data. This is a stylistic point, but you should consider condensing some of these explanations.
7. "On the other hand, systems such as Auto-Prognosis (Alaa & Schaar, 2018) focusses on longitudinal classification, but lack the simplicity of Scikit-learn and convenience of use out of the box that these systems are often lacking." <-- A few points here. i) I believe "focusses" should be "focus", since you're dealing with a plural (multiple "systems"); ii) No comma is needed after "classification"; iii) What makes tools like Auto-Prognosis difficult to use relative to your tool?
