### 2025-07-28 ce287a8
refactor(paper): typos & justification for Auto-Sklong [JOSS-Review]
