+++ paper/paper.md
@@ -26,26 +26,27 @@ Longitudinal data, characterised by repeated measurements of variables over time
 opportunities in
 machine learning. This paper introduces `Scikit-Longitudinal`, a Python library designed to address these challenges by
 providing a comprehensive set of tools for longitudinal data classification. Built to integrate with the
-popular `Scikit-Learn` library, `Scikit-Longitudinal` offers a robust solution for researchers and practitioners
+popular `Scikit-learn` library, `Scikit-Longitudinal` offers a robust solution for researchers and practitioners
 working with longitudinal datasets.
 
 # Summary
 
-`Scikit-Longitudinal` is an open-source Python library that enhances machine learning for longitudinal data
-classification and integrates with the `Scikit-Learn` environment [@pedregosa2011scikit].
+`Scikit-Longitudinal`, also abbreviated `Sklong`, is an open-source Python library that enhances machine learning for
+longitudinal data
+classification and integrates with the `Scikit-learn` environment [@pedregosa2011scikit].
 
 Longitudinal data, which consists of repeated measurements of variables across time points (referred to as
-`waves` [@ribeiro2019mini]), is extensively utilised in fields such as medicine, and social sciences. Unlike
-standard tabular datasets, it contains temporal relationships that necessitate specialised
+`waves` [@ribeiro2019mini]), is extensively utilised in fields such as medicine and social sciences. Unlike
+standard tabular datasets, longitudinal data contains temporal relationships that necessitate specialised
 processing [@kelloway2012longitudinal].
 
-`Scikit-Longitudinal` addresses this with a novel library that includes:
+`Sklong` addresses this with a novel library that includes:
 
 - **Data Preparation**: Utilities such as `LongitudinalDataset` for loading and structuring data, defining temporal
   feature groups, and other techniques.
 
 - **Data Transformation**: Methods to treat the temporal aspect of tabular data, by either (1) flattening the data
-  into a static representation (i.e. ignoring time indices) for standard machine learning to be performed (e.g.,
+  into a static representation (i.e., ignoring time indices) for standard machine learning to be performed (e.g.,
   `MarWavTimeMinus`, or `SepWav`),
   or (2) keeping the temporal structure (e.g., `MerWavTimePlus`), yet saving it for later use in longitudinal-data-aware
   `preprocessing` or `estimators` steps [@ribeiro2019mini].
@@ -64,13 +65,13 @@ processing [@kelloway2012longitudinal].
   ideally enhance classification performance.
 
 In total, the library implements 1 data preparation method, 4 data transformation methods, 1 preprocessing method, and
-6 estimators, 2 of which have been published as stand alone methods in the literature (the above named
+6 estimators, 2 of which have been published as stand-alone methods in the literature (the above named
 `LexicoRandomForestClassifier` and `NestedTreesClassifier` methods).
 
-`Scikit-Longitudinal` emphasises highly-typed, Pythonic code, with substantial test coverage (over 88%) and
+`Sklong` emphasises highly-typed, Pythonic code, with substantial test coverage (over 88%) and
 comprehensive documentation (over 72%).
 
-Finally, `Scikit-Longitudinal` can also be abbreviated `Sklong`, and is available
+Finally, `Sklong` is available
 on [PyPI](https://pypi.org/project/Scikit-longitudinal/). Feel free to explore the
 [official documentation](https://scikit-longitudinal.readthedocs.io/latest/) for various
 installation methods.
@@ -100,11 +101,11 @@ the goal is to predict the class label of an instance at a single time point (us
 
 There are two broad approaches for coping with longitudinal data [@ribeiro2019mini]: (1) **Data Transformation**:
 this approach involves preprocessing methods that convert longitudinal data into a standard, "flattened"
[…]
