### 2025-07-28 e87b5ed
refactor(paper): add missing dois when available [JOSS-Review]
