+++ paper/paper.bib
@@ -1,10 +1,12 @@
 @inproceedings{autoprognosis,
-  title={Autoprognosis: Automated clinical prognostic modeling via bayesian optimization with structured kernel learning},
-  author={Alaa, Ahmed and Schaar, Mihaela},
-  booktitle={Proceedings of the International Conference on Machine Learning},
-  pages={139--148},
-  year={2018},
-  organization={PMLR}
+  doi = {10.48550/ARXIV.1802.07207},
+  url = {https://arxiv.org/abs/1802.07207},
+  author = {Alaa,  Ahmed M. and van der Schaar,  Mihaela},
+  keywords = {Machine Learning (cs.LG),  Machine Learning (stat.ML),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
+  title = {AutoPrognosis: Automated Clinical Prognostic Modeling via Bayesian Optimization with Structured Kernel Learning},
+  publisher = {arXiv},
+  year = {2018},
+  copyright = {arXiv.org perpetual,  non-exclusive license}
 }
 
 @inproceedings{provost2024auto,
@@ -13,7 +15,8 @@
   booktitle={Proceedings of the 2024 IEEE International Conference on Bioinformatics and Biomedicine (BIBM)},
   pages={3645--3650},
   year={2024},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/BIBM62325.2024.10821737}
 }
 
 @article{pedregosa2011scikit,
@@ -39,7 +42,8 @@
   title={Mixed-effects models in S and S-PLUS},
   author={Pinheiro, Jos{\'e} and Bates, Douglas},
   year={2000},
-  publisher={Springer science \& business media}
+  publisher={Springer science \& business media},
+  doi = {10.1007/0-387-22747-4_8},
 }
 
 @article{GPBoost,
@@ -58,7 +62,8 @@
   booktitle={Proceedings of the 2017 IEEE International Conference on Data Mining Workshops (ICDMW)},
   pages={739--746},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icdmw.2017.102}
 }
 
 @mastersthesis{pomsuwan2018featureversion2,
@@ -75,6 +80,7 @@
   author={Ovchinnik, Sergey and Otero, Fernando and Freitas, Alex A},
   booktitle={Proceedings of the 37th ACM/SIGAPP Symposium on Applied Computing},
   pages={441--444},
+  doi = {10.1145/3477314.3507240},
   year={2022}
 }
 
@@ -95,6 +101,7 @@
   publisher={Springer Science and Business Media LLC},
   volume={57},
   pages={Article 84},
+  doi = {10.1007/s10462-024-10718-1},
   issn={1573-7462}
 }
 
@@ -102,7 +109,8 @@
   author={Li, Adam and Panda, Sambit and Xu, Haoyin and Ogihara, Itsuki},
   year={2024},
   license={PolyForm-Noncommercial-1.0.0},
-  title={{treeple: Modern decision-trees compatible with scikit-learn in Python.}}
+  title={{treeple: Modern decision-trees compatible with scikit-learn in Python.}},
+  doi = {10.5281/zenodo.14509519},
 }
 
 @article{theng2024feature,
@@ -113,6 +121,7 @@
   number={3},
   pages={1575--1637},
   year={2024},
+  doi = {10.1007/s10115-023-02010-5},
   publisher={Springer}
 }
 
