### 2025-07-28 #8481 @simonprovost
> Please address the reference issues listed above.
>
> ### Paper Comments:
>
> ...
> 7. "On the other hand, systems such as Auto-Prognosis (Alaa & Schaar, 2018) focusses on longitudinal classification, but lack the simplicity of Scikit-learn and convenience of use out of the box that these systems are often lacking." <-- A few points here. i) I believe "focusses" should be "focus", since you're dealing with a plural (multiple "systems"); ii) No comma is needed after "classification"; iii) What makes tools like Auto-Prognosis difficult to use relative to your tool?

- [x] Bib is now done. With the remaining "missing" DOIs being not found / not delivered.
- [x] All other comments been taken into account and changes been done accordingly. As for (7) extra justification been written down exceeding therefore the 1000 words paper to 1015, may this be "acceptable", @espottesmith ?

Hoping this is all good :)

Cheers,
