### 2025-07-28 #8481 @simonprovost
> Please address the reference issues listed above.
>
> ### Paper Comments:
>
> ...
> 7. "On the other hand, systems such as Auto-Prognosis (Alaa & Schaar, 2018) focusses on longitudinal classification, but lack the simplicity of Scikit-learn and convenience of use out of the box that these systems are often lacking." <-- A few points here. i) I believe "focusses" should be "focus", since you're dealing with a plural (multiple "systems"); ii) No comma is needed after "classification"; iii) What makes tools like Auto-Prognosis difficult to use relative to your tool?

Hi @espottesmith,

Thanks so much for such suggestions! As follows:

- [x] BIB is now done. With the remaining "missing" DOIs being not found / not delivered. Note one thing, `- 10.1109/icdmw.2017.102 may be a valid DOI for title: Feature Selection for the classification of longit...` is for the BIB referring to the master's thesis which has no DOI delivered.
- [x] All comments have been considered, and appropriate changes have been made.  As for (7), further justification has been written down, bringing the paper's total word count to 1015 (+15 words above). Is this "acceptable", @espottesmith?

Hoping this is all good :)

Cheers,
