### 2025-07-31 #8481 @simonprovost
I have no clues if I still could, I aim to solve some typos and remove one sentence in the paper following a read with my supervisor if that’s all right! Will do so by end of day (UK time zone)
