### 2025-07-31 6c12cb3
refactor(paper): typos & bib strengthening
