+++ paper/paper.md
@@ -40,7 +40,7 @@ Longitudinal data, which consists of repeated measurements of variables across t
 standard tabular datasets, longitudinal data contains temporal relationships that necessitate specialised
 processing [@kelloway2012longitudinal].
 
-`Sklong` addresses this with a novel library that includes:
+`Sklong` addresses this with a novel library that includes [@provost2024auto]:
 
 - **Data Preparation**: Utilities such as `LongitudinalDataset` for loading and structuring data, defining temporal
   feature groups, and other techniques.
@@ -114,13 +114,13 @@ classification.
 Standard Python libraries, such as `Scikit-learn` itself, lack support for longitudinal data, leading to inefficient and
 inaccurate analysis. `R` includes
 statistical packages for longitudinal data (e.g., `nlme` [@pinheiro2000mixed], `GPBoost` [@GPBoost]). However, they
-often are not suitable for machine learning workflows often created in Python. On the other side, systems like `Auto-Prognosis` [@autoprognosis] concentrate on longitudinal classification but do not
+often are not suitable for machine learning workflows often created in Python. On the other hand, systems like `Auto-Prognosis` [@autoprognosis] concentrate on longitudinal classification but do not
 have `Scikit-learn`'s ease of use. `Auto-Prognosis` encompasses more than just longitudinal machine learning, making it
 difficult to identify and investigate specific problems. Furthermore, it focuses on algorithm adaptation for prognosis
 rather than providing both data transformation and algorithm adaptation paths like Sklong, which limits user
 flexibility.
 
-Given the lack of Python libraries, integration with the popular `Scikit-learn` API, and
+Given the lack of Python libraries, lack of integration with the popular `Scikit-learn` API, and
 the absence of out-of-the-box solutions for longitudinal classification, there is a clear need for a library that
 provides tools for longitudinal data preparation, transformation, preprocessing, and estimation (model learning).
 
@@ -132,13 +132,11 @@ for regression or neural networks. Future development could expand the library i
 # Acknowledgements
 
 We thank the authors and contributors of `Scikit-learn` for their pioneering work in machine learning. We thank the
-NeuroData team for their contributions to `Scikit-Tree` [@Li_treeple_Modern_decision-trees], which enables modification
-of `Scikit-learn`'s Cython trees for optimisation performance. We are grateful to the researchers who contributed to the
+NeuroData team for their contributions to `Scikit-Tree` [@Li_treeple_Modern_decision_trees], which enables modification
+of `Scikit-learn`'s Cython trees for optimising performance. We are grateful to the researchers who contributed to the
 design of many primitives within `Sklong`, including Dr. Tossapol
 Pomsuwan [@pomsuwan2017feature; @pomsuwan2018featureversion2], Dr. Sergey Ovchinnik & Dr. Fernando
 Otero [@ovchinnik2022nested],
 and Dr. Caio Ribeiro [@ribeiro2019mini; @ribeiro2022new; @Ribeiro2024].
-
-This work was completed during a PhD project [@provost2024auto] at the University of Kent.
 
 # References
