+++ paper/paper.bib
@@ -72,7 +72,7 @@
   year={2018},
   month={Feb},
   type={Master by Research thesis},
-  school={School of Computing, University of Kent}
+  school={School of Computing, University of Kent, UK}
 }
 
 @inproceedings{ovchinnik2022nested,
@@ -88,7 +88,7 @@
   title={New Longitudinal Classification Approaches and Applications to Age-Related Disease Data},
   author={Ribeiro, Caio},
   year={2022},
-  school={School of Computing, University of Kent},
+  school={School of Computing, University of Kent, UK},
   type={Ph.D. Thesis}
 }
 
@@ -100,17 +100,31 @@
   journal={Artificial Intelligence Review},
   publisher={Springer Science and Business Media LLC},
   volume={57},
-  pages={Article 84},
+  pages={Article 84, 29 pages},
   doi = {10.1007/s10462-024-10718-1},
   issn={1573-7462}
 }
 
-@software{Li_treeple_Modern_decision-trees,
-  author={Li, Adam and Panda, Sambit and Xu, Haoyin and Ogihara, Itsuki},
-  year={2024},
-  license={PolyForm-Noncommercial-1.0.0},
-  title={{treeple: Modern decision-trees compatible with scikit-learn in Python.}},
-  doi = {10.5281/zenodo.14509519},
+@software{Li_treeple_Modern_decision_trees,
+  author       = {Li, Adam and
+                  Panda, Sambit and
+                  Xu, Haoyin and
+                  Ogihara, Itsuki},
+  title        = {treeple: Modern decision-trees compatible with
+                   scikit-learn in Python.
+                  },
+  month        = dec,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v0.10.0},
+  doi          = {10.5281/zenodo.14509519},
+  url          = {https://doi.org/10.5281/zenodo.14509519},
+  swhid        = {swh:1:dir:614a47a50342fc6d06c177227d3399e949d734c5
+                   ;origin=https://doi.org/10.5281/zenodo.8412279;vis
+                   it=swh:1:snp:ad9a1a1713eb219bf315cab6e3ac3e6f338c4
+                   ddd;anchor=swh:1:rel:71244bb3020c0c8921b79a025128b
+                   6210b047483;path=neurodata-treeple-ab12ca9
+                  },
 }
 
 @article{theng2024feature,
@@ -128,9 +142,8 @@
 @inproceedings{ribeiro2019mini,
   title={A mini-survey of supervised machine learning approaches for coping with ageing-related longitudinal datasets},
   author={Ribeiro, Caio and Freitas, Alex A},
-  booktitle={Proceedings of the 3rd Workshop on AI for Aging, Rehabilitation and Independent Assisted Living (ARIAL), held as part of IJCAI-2019},
-  year={2019},
-  note={5 pages}
+  booktitle={Proceedings of the 3rd Workshop on AI for Aging, Rehabilitation and Independent Assisted Living (ARIAL), held as part of IJCAI-2019, 5 pages},
+  year={2019}
 }
 
 @article{kotsiantis2007supervised,
