### 2025-07-31 #8481 @simonprovost
@espottesmith check references
