### 2025-07-31 #8481 @simonprovost
Hey @espottesmith  ,

A new version of the paper with 1004 words is ready; one sentence has been removed to avoid the 1015-ish words, plus typos have been altered, and the bibliography has been "fixed" in response to suggestions from my supervisor this morning.

Changes can be seen at: https://github.com/simonprovost/scikit-longitudinal/commit/6c12cb3627c368dcb4e51e8858ea73c69a281d0f

On top of things, as previously explained: DOIS and the rest have been checked and seem to be good.

Are we nearing the end:)?

Cheers,
