### 2025-08-01 020eeef
fix: setup.py & PyPI Links [cd build]
