+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "Scikit-longitudinal"
-version = "0.0.9"
+version = "0.0.10"
 description = "Scikit-longitudinal is an open-source Python library for longitudinal data analysis, building on Scikit-learn's foundation with tools tailored for repeated measures data."
 authors = [
     {name = "Provost Simon", email = "simon.gilbert.provost@gmail.com"},
@@ -60,11 +60,13 @@ dev = [
     "mkdocs-glightbox>=0.4.0",
 ]
 
-[urls]
+[project.urls]
 Homepage = "https://github.com/simonprovost/scikit-longitudinal"
 Documentation = "https://scikit-longitudinal.readthedocs.io/latest/"
-Tracker = "https://github.com/simonprovost/scikit-longitudinal/issues/"
-"JOSS paper" = "https://github.com/openjournals/joss-reviews/issues/8481"
+Releases = "https://github.com/simonprovost/scikit-longitudinal/releases"
+Repository = "https://github.com/simonprovost/scikit-longitudinal"
+Issues = "https://github.com/simonprovost/scikit-longitudinal/issues/"
+"Scientific Paper" = "https://github.com/openjournals/joss-reviews/issues/8481"
 
 [tool.uv]
 default-groups = []
