### 2025-08-01 #8481 @espottesmith
Hey @simonprovost,

I appreciate your work on the manuscript. I've looked over the reference, and I can see that there's nothing you can do (short of minting a new DOI for the master's thesis, I guess, which I'm not suggesting you do) to resolve the issue with the last remaining missing DOI. In other words, the references are fine now as-is.

The length of the manuscript is fine. 4 (or 15) words don't matter, as far as I'm concerned.
