### 2025-08-01 #8481 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
