### 2025-08-01 #8481 @simonprovost
Woops I openned the wrong PDF, too many version. Let me correct myself: Everything is fine with this version.

I also chcked the various hyperlinks:

- [x] ORCIDs
- [x] Review/Repo/Archive
- [x] Editor/Reviewers

Only the DOI is not working but I presume it is all right, needs to wait to be published.

Cheers
