### 2025-08-01 767ed3b
refactor(paper): citations & letter-capitalisation [JOSS-Review]
