+++ paper/paper.bib
@@ -3,7 +3,7 @@
   url = {https://arxiv.org/abs/1802.07207},
   author = {Alaa,  Ahmed M. and van der Schaar,  Mihaela},
   keywords = {Machine Learning (cs.LG),  Machine Learning (stat.ML),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
-  title = {AutoPrognosis: Automated Clinical Prognostic Modeling via Bayesian Optimization with Structured Kernel Learning},
+  title = {AutoPrognosis: {A}utomated {C}linical {P}rognostic {M}odeling via {B}ayesian {O}ptimization with {S}tructured {K}ernel {L}earning},
   publisher = {arXiv},
   year = {2018},
   copyright = {arXiv.org perpetual,  non-exclusive license}
@@ -20,7 +20,7 @@
 }
 
 @article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
+  title={Scikit-learn: Machine learning in {P}ython},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={the Journal of machine Learning research},
   volume={12},
@@ -111,7 +111,7 @@
                   Xu, Haoyin and
                   Ogihara, Itsuki},
   title        = {treeple: Modern decision-trees compatible with
-                   scikit-learn in Python.
+                   scikit-learn in {P}ython.
                   },
   month        = dec,
   year         = 2024,
