### 2025-08-01 #8481 @simonprovost
👋 all,

I would like to express my heartfelt gratitude to @blengerich, @TahiriNadia, and @espottesmith for (1) always so swiftly responding to me, and I am taking the opportunity to profoundly apologising for the sometimes (most of the time... so sorry!) brutal daily tagging for friendly reminders; (2) your dedication to Scikit-Longitudinal and great comments and suggestions; and (3) giving me motivation to help with peer reviewing @ JOSS! I would like to give back the positive energy I got to others and empower our OSS world :)

I would also like to emphasise that your many iterations of help have not gone forgotten, so if you require my help with anything, please tag me in any thread or any GitHub issues, or simply pop me an email; I am more than happy to help!

Wishing everyone a, hopefully, restful weekend ahead!

On a last note, many thanks to @crvernon for tying everything up!

Cheers 🙏
