### 2025-08-01 #8481 @crvernon
@simonprovost yes the DOI will resolve.  Thank you.
